refactor(demo): modularize demo data generators into separate files
Split the monolithic demo_data.dart into focused generators for each domain model (tags, habits, tasks, notes, app usages) to improve code organization and maintainability. Each generator now handles its specific data creation logic with appropriate separation of concerns.
