+++ src/lib/core/application/features/demo/models/demo_data/generators/generators.dart
@@ -1,0 +1,8 @@
+/// Demo data generators barrel file
+library;
+
+export 'demo_tags.dart';
+export 'demo_habits.dart';
+export 'demo_tasks.dart';
+export 'demo_notes.dart';
+export 'demo_app_usages.dart';
